% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ama_interviews}
\alias{ama_interviews}
\title{Extract jobs that will have AMA Interviews}
\usage{
ama_interviews(jobs_data)
}
\arguments{
\item{jobs_data}{A data.frame of jobs from which to filter}
}
\value{
A data.frame of jobs that will have AMA interviews
}
\description{
Extract jobs that will have AMA Interviews
}
\examples{
jobs_data <- get_job_listings()
ama <- ama_interviews(jobs_data)
}
