% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{format_jobs}
\alias{format_jobs}
\title{Format Jobs Data}
\usage{
format_jobs(html_text)
}
\description{
Cleans up the text output of the HTML table imported from
marketingphdjobs.org
}
\details{
End result is a tidy data frame
}
