% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_job_listings}
\alias{get_job_listings}
\title{Get Job Listings from marketingphdjobs.org}
\usage{
get_job_listings()
}
\value{
Returns an object of class "data.frame"
}
\description{
Extract job listings from the website marketingphdjobs.org
}
\examples{
job_data <- get_job_listings()
}
