% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_new_jobs}
\alias{get_new_jobs}
\title{Find new jobs based on a date criterion}
\usage{
get_new_jobs(date_lower, date_upper = NULL)
}
\arguments{
\item{date_lower}{A date from which to start the search}

\item{date_upper}{A date from which the search ends}
}
\value{
A data.frame of jobs posted after date_lower and before date_upper
}
\description{
Find new jobs based on a date criterion
}
\examples{
## between two set dates
date_start <- lubridate::ymd("2019/06/18")
date_end   <- lubridate::ymd("2019/06/20")
new_jobs <- get_new_jobs(date_start, date_end)

## all jobs from a given date until today
new_jobs2 <- get_new_jobs(date_lower = date_start)
}
