% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{job_type}
\alias{job_type}
\title{Extract jobs of a certain position}
\usage{
job_type(jobs_data, job_level)
}
\arguments{
\item{jobs_data}{A data.frame of jobs from which to filter}

\item{job_level}{A character string containing the job level to locate}
}
\value{
A data.frame of jobs of a certain level
}
\description{
Extract jobs of a certain position
}
\examples{
jobs_data <- get_job_listings()
assist_prof <- job_type(jobs_data, "assistant")
## equivalent to
assist_prof2 <- job_type(jobs_data, "Assistant")
}
