% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{jobs_in_country}
\alias{jobs_in_country}
\title{Extract job listings for a certain country}
\usage{
jobs_in_country(jobs_data, job_location)
}
\arguments{
\item{jobs_data}{A data.frame of jobs from which to filter}

\item{job_level}{A character string containing the country of interest}
}
\value{
A data.frame of job postings in a certain country
}
\description{
Extract job listings for a certain country
}
\examples{
df <- get_job_listings()
us_jobs <- jobs_in_country(df, "united states")
## Multiple countries
selected_jobs <- jobs_in_country(df, "united states|canada")
}
